interfaceの統一とフォルダ整理
