+++ src/App.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import { Routes, Route } from 'react-router-dom';
-import type { ModalState } from './types';
+import type { ModalState } from './interface/types';
 
 // ページコンポーネントをインポート
 import HomePage from './pages/HomePage';
