+++ src/interface/auth.ts
@@ -1,0 +1,11 @@
+export interface LoginValues {
+  email: string;
+  password: string;
+}
+
+export interface RegisterValues {
+  name: string;
+  email: string;
+  password: string;
+  passwordConfirm: string;
+}
