--- src/types.ts
@@ -1,11 +1,0 @@
-export type ActiveTab = 'login' | 'register' | 'forgot';
-
-export interface ModalState {
-  isOpen: boolean;
-  tab: ActiveTab;
-}
-
-export interface User{
-  id: number;
-  username: string;
-}
